+++ pkg/frontend/transport/retry_test.go
@@ -8,6 +8,8 @@ import (
 	"github.com/thanos-io/thanos/pkg/pool"
 	"github.com/weaveworks/common/httpgrpc"
 	"go.uber.org/atomic"
+
+	"github.com/cortexproject/cortex/pkg/storegateway"
 )
 
 func TestRetry(t *testing.T) {
@@ -41,6 +43,28 @@ func TestNoRetryOnChunkPoolExhaustion(t *testing.T) {
 			return &httpgrpc.HTTPResponse{
 				Code: 500,
 				Body: []byte(pool.ErrPoolExhausted.Error()),
+			}, nil
+		}
+		return &httpgrpc.HTTPResponse{
+			Code: 200,
+		}, nil
+
+	})
+
+	require.NoError(t, err)
+	require.Equal(t, int32(500), res.Code)
+}
+
+func TestNoRetryOnMaxConcurrentBytesLimitExceeded(t *testing.T) {
+	tries := atomic.NewInt64(3)
+	r := NewRetry(3, nil)
+	ctx := context.Background()
+	res, err := r.Do(ctx, func() (*httpgrpc.HTTPResponse, error) {
+		try := tries.Dec()
+		if try > 1 {
+			return &httpgrpc.HTTPResponse{
+				Code: 500,
+				Body: []byte(storegateway.ErrMaxConcurrentBytesLimitExceeded.Error()),
 			}, nil
 		}
 		return &httpgrpc.HTTPResponse{